refactor: move wings compatibility code under a feature flag
